Modifying error message

+++ src/layers/services/room.service.js
@@ -1,3 +1,4 @@
+const { IoTSecureTunneling } = require('aws-sdk');
 const RoomRepository = require('../repositories/room.repository');
 
 class RoomService {
@@ -7,11 +8,15 @@ class RoomService {
         this.roomRepository = new RoomRepository();
     }
 
+    errResponse = (status, message) => {
+        return { status, success: false, message };
+    };
+
     postRoom = async (ownerUserId, roomName, content, category) => {
         return await this.roomRepository.postRoom(ownerUserId, roomName, content, category);
     };
 
-    getRoom = async (req, res, next) => {
+    getRoom = async () => {
         return await this.roomRepository.getRoom();
     };
 
@@ -20,7 +25,13 @@ class RoomService {
     };
 
     deleteRoom = async (roomId) => {
-        return await this.roomRepository.deleteRoom(roomId);
+        const roomInfo = await this.roomRepository.getOneRoom(roomId);
+        if (!roomInfo) return this.errResponse(400, '존재하지 않는 룸입니다.');
+
+        const deleteRoom = await this.roomRepository.deleteRoom(roomId);
+        if (!deleteRoom) return this.errResponse(401, '알 수 없는 에러');
+
+        return { status: 200, success: true, result: {} };
     };
 }
 module.exports = RoomService;
